+++ ppci/ir.py
@@ -152,10 +152,10 @@ u8 = UnsignedIntegerTyp("u8", 8)  #: Unsigned 8-bit type
 ptr = PointerTyp("ptr")  #: Pointer type
 
 # Atalla specific types
-f16 = FloatingPointTyp("f16", 16)  #: 16-bit floating point type
+bf16 = FloatingPointTyp("bf16", 16)  #: 16-bit floating point type
 vec = VectorTyp("vec", 32*16)  #: Vector type
 
-value_types = [vec, f16, f32, f64, i64, i32, i16, i8, u64, u32, u16, u8]
+value_types = [vec, bf16, f32, f64, i64, i32, i16, i8, u64, u32, u16, u8]
 all_types = value_types + [ptr]
 type_name_map = {t.name.lower(): t for t in value_types}
